+++ src/DataFilters/Grammar/Syntax/ContainsExpression.cs
@@ -5,11 +5,7 @@ namespace DataFilters.Grammar.Syntax
     /// <summary>
     /// A <see cref="FilterExpression"/> that holds a string value
     /// </summary>
-#if NETSTANDARD1_3
     public sealed class ContainsExpression : FilterExpression, IEquatable<ContainsExpression>
-#else
-    public record ContainsExpression : FilterExpression, IEquatable<ContainsExpression>
-#endif
     {
         /// <summary>
         /// The value that was between two <see cref="AsteriskExpression"/>
@@ -40,7 +36,6 @@ namespace DataFilters.Grammar.Syntax
         ///<inheritdoc/>
         public override double Complexity => 1.5;
 
-#if NETSTANDARD1_3
         ///<inheritdoc/>
         public override bool Equals(object obj) => Equals(obj as ContainsExpression);
 
@@ -52,7 +47,5 @@ namespace DataFilters.Grammar.Syntax
 
         ///<inheritdoc/>
         public override string ToString() => this.Jsonify();
-
-#endif
     }
 }