+++ src/DataFilters/Grammar/Syntax/DateTimeExpression.cs
@@ -5,11 +5,7 @@ namespace DataFilters.Grammar.Syntax
     /// <summary>
     /// A <see cref="FilterExpression"/> implementation that can holds a datetime value
     /// </summary>
-#if NETSTANDARD1_3
     public sealed class DateTimeExpression : FilterExpression, IEquatable<DateTimeExpression>, IBoundaryExpression
-#else
-    public record DateTimeExpression : FilterExpression, IEquatable<DateTimeExpression>, IBoundaryExpression
-#endif
     {
         /// <summary>
         /// Date part of the expression
@@ -79,7 +75,6 @@ namespace DataFilters.Grammar.Syntax
             time = Time;
         }
 
-#if NETSTANDARD1_3
         /// <inheritdoc/>
         public bool Equals(DateTimeExpression other) => (Date, Time).Equals((other?.Date, other?.Time));
 
@@ -91,6 +86,5 @@ namespace DataFilters.Grammar.Syntax
 
         ///<inheritdoc/>
         public override string ToString() => this.Jsonify();
-#endif 
     }
 }