+++ src/DataFilters/Grammar/Syntax/FilterExpression.cs
@@ -6,26 +6,6 @@
     /// <para>
     /// <see cref="FilterExpression"/>s can take many forms : from the simplest <see cref="ConstantValueExpression"/> to more complex <see cref="GroupExpression"/>s.
     /// </para>
-#if NETSTANDARD1_4_OR_GREATER || NET5_0_OR_GREATER
-    public abstract record FilterExpression : IHaveComplexity
-    {
-        /// <summary>
-        /// Tests if <paramref name="other"/> is equivalent to the current instance.
-        /// </summary>
-        /// <param name="other"><see cref="FilterExpression"/> against which the current instance will test is equivalency.</param>
-        /// <remarks>
-        /// The default implementation defers to <see cref="object.Equals(object)"/> implementation.
-        /// The meaning of the equivalency of two <see cref="FilterExpression"/>s is left to the implementor.
-        /// </remarks>
-        /// <returns>
-        /// <c>true</c> if <paramref name="other"/> is "equivalent" to the current instance.
-        /// </returns>
-        public virtual bool IsEquivalentTo(FilterExpression other) => Equals(other);
-
-        ///<inheritdoc/>
-        public virtual double Complexity => 1;
-    } 
-#else
     public abstract class FilterExpression : IHaveComplexity
     {
         /// <summary>
@@ -44,5 +24,4 @@
         ///<inheritdoc/>
         public virtual double Complexity => 1;
     }
-#endif
 }