+++ src/Datafilters.Expressions/FilterExtensions.cs
@@ -1,12 +1,9 @@
-﻿using DataFilters.Converters;
-using DataFilters.Grammar.Exceptions;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
-using System.Text.RegularExpressions;
+
 using static DataFilters.FilterOperator;
 using static System.Linq.Expressions.Expression;